+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2025-09-05
 ### Added
  - Introduce clearance table generation and clearance automatic table update from kicad project
 
@@ -36,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial release 
 
-[unreleased]: https://github.com/upb-lea/KiClearance/compare/0.1.2...HEAD
+[unreleased]: https://github.com/upb-lea/KiClearance/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/upb-lea/KiClearance/compare/0.1.0...0.2.0
 [0.1.2]: https://github.com/upb-lea/KiClearance/compare/0.1.0...0.1.2
 [0.1.1]: https://github.com/upb-lea/KiClearance/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/upb-lea/KiClearance/compare/0.0.2...0.1.0